refactor(portfolio): extract hooks
